rename metrics lib
