+++ cli/cli/commands/service/inspect/inspect.go
@@ -22,7 +22,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis/cli/cli/helpers/user_services"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/out"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface"
-	metrics_client "github.com/kurtosis-tech/kurtosis/metrics-library/golang/lib/client"
+	"github.com/kurtosis-tech/kurtosis/metrics-library/golang/lib/metrics_client"
 	"github.com/kurtosis-tech/stacktrace"
 )
