+++ core/server/api_container/server/api_container_service.go
@@ -22,7 +22,7 @@ import (
 	"unicode"
 
 	"github.com/kurtosis-tech/kurtosis/core/server/commons/yaml_parser"
-	metrics_client "github.com/kurtosis-tech/kurtosis/metrics-library/golang/lib/client"
+	"github.com/kurtosis-tech/kurtosis/metrics-library/golang/lib/metrics_client"
 
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/uuid_generator"
