+++ metrics-library/golang/lib/metrics_client/is_ci_test.go
@@ -1,4 +1,4 @@
-package client
+package metrics_client
 
 import (
 	"github.com/stretchr/testify/require"
