+++ metrics-library/golang/lib/metrics_client/metrics_client.go
@@ -1,4 +1,4 @@
-package client
+package metrics_client
 
 type MetricsClient interface {
 	TrackShouldSendMetricsUserElection(didUserAcceptSendingMetrics bool) error
